Add toggling ability
