+++ src/types.ts
@@ -1,0 +1,6 @@
+export interface BookEntry {
+  id: number;
+  title: string;
+  count: number;
+  price: number;
+}
